freetype to Revision: 3f3e3de34ee3613d621b643c58a40b93148e0932 Author: Honnesh Ramachandra <hramachandra@adobe.com> Date: 06/11/2024 6:28:41 Message: * src/cff/cffobjs.c (cff_face_init): Better handling of Type0 fonts.
This issue relates to the PDF specification and its usage of CFF-based
embedded Type0 fonts.  For Type0 fonts containing CID-keyed descendant
CFF-based fonts, the glyph index is the CID present in the content stream.
As a result, there is no requirement for a 'cmap' table to be present in the
font since the glyph indices are derived directly from the CIDs.  FreeType
throws an error when it doesn’t find a 'cmap' table for such an Open Type
font containing CFF outlines.  This commit relaxes this requirement for a
'cmap' table for such fonts.
----
Modified: src/cff/cffobjs.c